requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
distlib==0.3.1
Django==3.0.8
django-ckeditor==6.0.0
django-js-asset==1.2.2
filelock==3.0.12
gunicorn==20.0.4
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==8.0.0
pylint==2.6.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.30
whitenoise==5.2.0
wrapt==1.12.1
